Security fix: Use different keys for each token.
This fixes a bug which allowed attackers who could
read the cookie token to generate a form token. It
does not allow classical CSRF attacks, because the
attacker still needs to be able to read the cookie
(this is only a hole if the attacker can sniff the
cookie token, but not the form token in the body).

This version will reject form tokens from previous
versions; users will need to refresh AJAX pages to
get a new valid form token.